refactor: extract inline tests from error.ts to separate error.test.ts
- Move all chainErr and ErrorArray tests to dedicated test file
- Remove problematic import.meta.vitest usage that was causing typecheck failures
- Fix torus-worker typecheck errors
- Maintain full test coverage with proper Vitest imports
